pyproject.toml
[project]
name = "cw7-habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduard Slobodyanik",email = "slobodyanik.ed@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "django (==4.2.2)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "ipython (>=8.32.0,<9.0.0)",
    "coverage (>=7.6.12,<8.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
    "django-phonenumber-field (>=8.0.0,<9.0.0)",
    "phonenumbers (>=8.13.55,<9.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "crispy-bootstrap4 (>=2024.10,<2025.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "telebot (>=0.0.5,<0.0.6)",

]

[tool.coverage.run]
omit = [
    "*/migrations/*",  # Исключаем все файлы миграций
    "*/__init__.py",   # Исключаем все служебные __init__.py
    "manage.py",       # Исключаем manage.py
    "config/asgi.py",  # Исключаем служебные файлы проекта
    "config/wsgi.py",
]




[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
